Fix: environment variables from Procfile were dropped
When SimpleSchema validated the process specification
for insertion into a Process document, it removed fields from the
process specification object. The insertion into a ProcfileEntry
didn't even get the field. This is resolved by passing a deep-cloned
object to Process.

Then, once it got the field, the validation for ProcfileEntry also
stripped all fields, because they weren't in the schema for
ProcfileEntry.env. Solution is blackbox: true.